Bug Fix
Change of test does not change its description in terms and conditions Window_Loaded

Enhancements:
Allow Floating point scores for tests
Change of email addresses for web

+++ AMDES_KBS/AMDES_KBS/Controllers/CLIPSController.cs
@@ -764,8 +764,8 @@ namespace AMDES_KBS.Controllers
                             int id = int.Parse(fav.GetFactSlot("GroupID").ToString().Remove(0, 1));
                             if (id == y)
                             {
-                                int succArg = int.Parse(fav.GetFactSlot("SuccessArg").ToString());
-                                int trueCount = int.Parse(fav.GetFactSlot("TrueCount").ToString());
+                                float succArg = float.Parse(fav.GetFactSlot("SuccessArg").ToString());
+                                float trueCount = float.Parse(fav.GetFactSlot("TrueCount").ToString());
                                 //20151012 - Bug here should not be tagged to individual symptom
                                 // s.SymptomName += "Patient's Score: " + trueCount + ", Normal Score: " + succArg + " or more";
                                 s.ScoreString = "Patient's Score: " + trueCount + ", Normal Score: " + succArg + " or more";
@@ -807,8 +807,8 @@ namespace AMDES_KBS.Controllers
                             int id = int.Parse(fav.GetFactSlot("GroupID").ToString().Remove(0, 1));
                             if (id == y)
                             {
-                                int succArg = int.Parse(fav.GetFactSlot("SuccessArg").ToString());
-                                int trueCount = int.Parse(fav.GetFactSlot("TrueCount").ToString());
+                                float succArg = float.Parse(fav.GetFactSlot("SuccessArg").ToString());
+                                float trueCount = float.Parse(fav.GetFactSlot("TrueCount").ToString());
                                 //20151012 - Bug here should not be tagged to individual symptom
                                 //s.SymptomName += "- Patient's Score: " + trueCount + ", Normal Score: " + succArg + " or more";
                                 s.ScoreString = "Patient's Score: " + trueCount + ", Normal Score: " + succArg + " or more";
